pyproject.toml
[tool.poetry]
name = "url_shorting"
version = "0.1.0"
description = ""
authors = ["Michael <edd4599@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^2.0.3"
Flask-SQLAlchemy = "^2.5.1"
Flask-Script = "^2.0.6"
Flask-Migrate = "^3.1.0"
psycopg2-binary = "^2.9.3"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
